+++ src/backend/trn/op/target_builtin.cc
@@ -23,6 +23,8 @@
  *
  *  builtin intrinsic operators specific to Trainium target.
  */
+#include <tvm/ffi/function.h>
+#include <tvm/runtime/base.h>
 #include <tvm/tirx/builtin.h>
 #include <tvm/tirx/op.h>
 #include <tvm/tirx/op_attr_types.h>
@@ -160,6 +162,8 @@ void RegisterNKIIntrinsicAliases() {
 }  // namespace
 
 #undef TIRX_DEFINE_BUILTIN_FUNC
+
+TVM_FFI_STATIC_INIT_BLOCK() { RegisterTRNTargetBuiltins(); }
 
 }  // namespace builtin
 }  // namespace tirx
