+++ src/backend/webgpu/codegen/target_kind.cc
@@ -18,50 +18,63 @@
  */
 
 /*!
- * \file register.cc
- * \brief OpenCL compiler backend static registration.
+ * \file target_kind.cc
+ * \brief WebGPU compiler backend static registration.
  */
 #include <dlpack/dlpack.h>
 #include <tvm/ffi/function.h>
+#include <tvm/ir/expr.h>
 #include <tvm/runtime/base.h>
 #include <tvm/target/target.h>
 #include <tvm/target/target_kind.h>
 
 namespace tvm {
 namespace backend {
-namespace opencl {
+namespace webgpu {
+
+ffi::Map<ffi::String, ffi::Any> UpdateWebGPUAttrs(ffi::Map<ffi::String, ffi::Any> target) {
+  bool subgroups = false;
+  if (target.count("supports_subgroups")) {
+    subgroups = Downcast<IntImm>(target.at("supports_subgroups"))->value != 0;
+  }
+
+  if (target.count("thread_warp_size")) {
+    int64_t thread_warp_size = Downcast<IntImm>(target.at("thread_warp_size"))->value;
+    TVM_FFI_ICHECK(subgroups || thread_warp_size <= 1)
+        << "WebGPU target with thread_warp_size=" << thread_warp_size
+        << " requires supports_subgroups=true";
+  }
+
+  if (subgroups) {
+    target.Set("thread_warp_size", int64_t(32));
+  }
+  return target;
+}
 
 void RegisterTargetKind() {
   namespace refl = tvm::ffi::reflection;
 
-  TVM_REGISTER_TARGET_KIND("opencl", kDLOpenCL)
-      .add_attr_option<int64_t>("max_threads_per_block", refl::DefaultValue(256))
-      .add_attr_option<int64_t>("max_shared_memory_per_block", refl::DefaultValue(16384))
+  TVM_REGISTER_TARGET_KIND("webgpu", kDLWebGPU)
       .add_attr_option<int64_t>("max_num_threads", refl::DefaultValue(256))
+      .add_attr_option<bool>("supports_subgroups", refl::DefaultValue(false))
       .add_attr_option<int64_t>("thread_warp_size", refl::DefaultValue(1))
-      .add_attr_option<int64_t>("texture_spatial_limit", refl::DefaultValue(16384))
-      .add_attr_option<int64_t>("texture_depth_limit", refl::DefaultValue(2048))
-      // Qualcomm OpenCL runtimes may crash when the number of kernel arguments is too large.
-      .add_attr_option<int64_t>("max_function_args", refl::DefaultValue(128))
-      .add_attr_option<int64_t>("image_base_address_alignment", refl::DefaultValue(64))
-      .set_default_keys({"opencl", "gpu"});
+      .set_target_canonicalizer(UpdateWebGPUAttrs)
+      .set_default_keys({"webgpu", "gpu"});
 }
 
-}  // namespace opencl
+}  // namespace webgpu
 }  // namespace backend
 
 namespace codegen {
-void RegisterOpenCLCodegen();
-void RegisterOpenCLDeviceScopeCompatibility();
+void RegisterWebGPUCodegen();
 namespace intrin {
-void RegisterOpenCLIntrinRules();
+void RegisterWebGPUIntrinRules();
 }  // namespace intrin
 }  // namespace codegen
 }  // namespace tvm
 
 TVM_FFI_STATIC_INIT_BLOCK() {
-  tvm::backend::opencl::RegisterTargetKind();
-  tvm::codegen::intrin::RegisterOpenCLIntrinRules();
-  tvm::codegen::RegisterOpenCLCodegen();
-  tvm::codegen::RegisterOpenCLDeviceScopeCompatibility();
+  tvm::backend::webgpu::RegisterTargetKind();
+  tvm::codegen::intrin::RegisterWebGPUIntrinRules();
+  tvm::codegen::RegisterWebGPUCodegen();
 }
